Lifted datepicker outside of component for better state management
